check if columns already exists

+++ database/migrations/2022_03_03_115304_alter_accommodations_add_created_by.php
@@ -14,9 +14,11 @@ class AlterAccommodationsAddCreatedBy extends Migration
     public function up()
     {
         Schema::table('accommodations', function (Blueprint $table) {
-            $table->integer('created_by')
-                ->after('is_free')
-                ->nullable();
+            if (!Schema::hasColumn('accommodations', 'created_by')) {
+                $table->integer('created_by')
+                    ->after('is_free')
+                    ->nullable();
+            }
         });
     }
 
@@ -28,7 +30,9 @@ class AlterAccommodationsAddCreatedBy extends Migration
     public function down()
     {
         Schema::table('accommodations', function (Blueprint $table) {
-            $table->dropColumn('created_by');
+            if (Schema::hasColumn('accommodations', 'created_by')) {
+                $table->dropColumn('created_by');
+            }
         });
     }
 }
